use LocationStreamingService instead of LocationBackgroundService
